Separate main into layers
